requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.2.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
flask>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
joblib>=1.1.0
requests>=2.28.0
jupyter>=1.0.0
streamlit>=1.28.0